+++ modules/settings.py
@@ -34,7 +34,7 @@ DEFAULT_SETTINGS = {
         "password": "",
     },
     "theme": "dark",
-    "version": "3.9",
+    "version": "4.0",
 }